+++ app/layout.js
@@ -1,27 +1,15 @@
-import "@rainbow-me/rainbowkit/styles.css";
+import '@rainbow-me/rainbowkit/styles.css';
+import { Providers } from './providers';
 
-import { getDefaultWallets, RainbowKitProvider } from "@rainbow-me/rainbowkit";
-import { createConfig, WagmiProvider } from "wagmi";
-import { sepolia } from "wagmi/chains";
-
-const { connectors } = getDefaultWallets({
-  appName: "CryptoDevs DAO",
-  projectId: "6ac21ee292826c2af6047b3e43e14054",
-});
-
-const config = createConfig({
-  chains: [sepolia],
-  autoConnect: true,
-  connectors,
-  publicClient,
-});
-
-export default function App({ Component, pageProps }) {
+function RootLayout({ children }) {
   return (
-    <WagmiProvider config={config}>
-      <RainbowKitProvider chains={chains}>
-        <Component {...pageProps} />
-      </RainbowKitProvider>
-    </WagmiProvider>
+    <html lang="en">
+      <body>
+        <Providers>{children}</Providers>
+      </body>
+    </html>
+    
   );
 }
+
+export default RootLayout;
