+++ package.json
@@ -9,8 +9,18 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@rainbow-me/rainbowkit": "^2.0.2",
+    "@tanstack/react-query": "^5.28.4",
+    "next": "14.1.3",
+    "pino-pretty": "^10.3.1",
     "react": "^18",
     "react-dom": "^18",
-    "next": "14.1.3"
+    "viem": "^2.8.12",
+    "wagmi": "^2.5.11"
+  },
+  "devDependencies": {
+    "@types/react": "18.2.67",
+    "encoding": "^0.1.13",
+    "typescript": "5.4.2"
   }
 }
